agent-eval/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "steam-agent-eval"
version = "0.1.0"
description = "Evaluation harness for Steam analytics conversational agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pytest>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=0.21",
    "rich>=13.0",
    "tabulate>=0.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
